requirements.txt
cryptography
urllib3
paho-mqtt